Extract message tool-call/noise analysis into message_analysis.py (WS5 seam)
Move the pure message-list inspection helpers out of engine.py into a new
message_analysis.py module: _tool_call_id, _assistant_tool_call_ids,
_matched_tool_call_ids, and _is_synthetic_assistant_noise (with its
_SYNTHETIC_ASSISTANT_NOISE vocabulary). engine.py imports the four it calls;
the noise vocabulary stays internal to the module.

Behaviour-preserving refactor (WS5 engine decomposition). The moved code is
byte-identical to its pre-move form. Rollback: git revert this commit restores
the inline block; no data or state migration involved.
